Move arguments-parsing logic to a utility class
