+++ internal-api/src/main/java/datadog/trace/bootstrap/config/provider/AgentArgsInjector.java
@@ -11,7 +11,7 @@ public class AgentArgsInjector {
    * @param agentArgs Agent arguments to be parsed and set
    */
   public static void injectAgentArgsConfig(String agentArgs) {
-    Map<String, String> args = AgentArgsParser.parseAgentArgs(agentArgs);
+    Map<String, String> args = Strings.parseCommaSeparatedKeyValueList(agentArgs);
     injectAgentArgsConfig(args);
   }
